+++ hail/python/hailtop/config/__init__.py
@@ -1,7 +1,7 @@
 from .user_config import (get_user_config, get_user_config_path,
                           get_remote_tmpdir, configuration_of)
 from .deploy_config import get_deploy_config, DeployConfig
-from .variables import ConfigVariable, config_variables
+from .variables import ConfigVariable
 
 __all__ = [
     'get_deploy_config',
@@ -11,5 +11,4 @@ __all__ = [
     'DeployConfig',
     'ConfigVariable',
     'configuration_of',
-    'config_variables',
 ]
